+++ pybettor/bet_calc.py
@@ -2,28 +2,25 @@ from fractions import Fraction
 from .convert_odds import convert_odds
 
 
-def bet_calc(risk, odds, category: str = "us"):
+def bet_calc(risk: int or float, odds: int or float, category: str = "us") -> float:
     """Calculate Bet Payout
     This function calculates the payout for a given bet.
 
     Args:
-        risk (float): Unit size of your bankroll, typically 1% of bankroll (100)
-        odds (float): Odds or Implied Win Probability for the bet
-        category (str, optional): 'us', 'dec', 'frac', 'prob'
-            Odds category.
-            Defaults to "us".
+        risk (int, float): Unit size of your bankroll, typically 1% of bankroll (100)
+        odds (int, float): Odds or Implied Win Probability for the bet
+        category (str, optional): type of odds. Defaults to "us". \n
+            'us', American Odds \n
+            'dec', Decimal Odds \n
+            'frac', Fractional Odds \n
+            'prob', Implied Probability
 
     Returns:
-        List[float]: Payout of a bet
+        float: Payout of a bet
     """
 
-    if type(risk) is not list:
-        risk = [risk]
-
-    if type(odds) is not list:
-        odds = [odds]
-
-    assert all(isinstance(x, (int, float)) for x in odds), "odds must be numeric"
+    assert isinstance(risk, (int, float)), "risk must be numeric"
+    assert isinstance(odds, (int, float)), "odds must be numeric"
     assert category in [
         "us",
         "frac",
@@ -34,13 +31,9 @@ def bet_calc(risk, odds, category: str = "us"):
     if category == "dec":
         pass
     else:
-        odds = convert_odds(odds, cat_in=category, cat_out="dec")
+        odds = convert_odds(odds, cat_in=category, cat_out="dec")[0]
 
     bets = {"odds": odds, "risk": risk}
-    payout = []
-    for i in range(len(bets["odds"])):
-        # the magic
-        payout_temp = round(bets["risk"][i] * bets["odds"][i], 2)
-        payout.append(payout_temp)
+    payout = round(bets["risk"] * bets["odds"], 2)
 
     return payout