+++ pybettor/parlay_calc.py
@@ -2,7 +2,7 @@ import numpy as np
 from .convert_odds import convert_odds
 
 
-def parlay_calc(risk: float = 100.0, odds=[-110, -110], category: str = "us"):
+def parlay_calc(risk: float = 100.0, odds=[-110, -110], category: str = "us") -> float:
     """Parlay Calculation
     This function calculates the payout of parlay bets.