+++ pybettor/true_implied_prob.py
@@ -1,9 +1,8 @@
 from .implied_prob import implied_prob
 from .true_probability import true_probability
-import pandas as pd
 
 
-def true_implied_prob(line: float, odds, category: str = "us"):
+def true_implied_prob(line: float, odds, category: str = "us") -> dict:
     """Bet Implied Probability
     This function calculates the implied probability of a bet given the odds.
 
@@ -12,7 +11,7 @@ def true_implied_prob(line: float, odds, category: str = "us"):
         category (str, optional): type of odds. Defaults to "us".
 
     Returns:
-        float: probability
+        dict: true_prob, imp_prob
     """
 
     assert isinstance(line, (int, float)), "line much be numeric"
@@ -26,8 +25,9 @@ def true_implied_prob(line: float, odds, category: str = "us"):
     true_prob = true_probability(line=line, odds=odds, category=category)
     imp_prob = implied_prob(odds=line, category=category)
 
-    df = pd.DataFrame(
-        list(zip([true_prob], imp_prob)), columns=["true_prob", "imp_prob"]
-    )
+    mydict = {
+        "true_prob": true_prob,
+        "imp_prob": imp_prob[0],
+    }
 
-    return df
+    return mydict